Split segment.py into some modules.
